requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
appnope==0.1.3
asttokens==2.2.1
autopep8==2.0.2
backcall==0.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coloredlogs==15.0.1
comm==0.1.3
debugpy==1.6.7
decorator==5.1.1
docopt==0.6.2
executing==1.2.0
filelock==3.11.0
Flask==2.2.3
flatbuffers==23.3.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
graphviz==0.8.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.3.0
ipykernel==6.22.0
ipython==8.12.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
mxnet==1.6.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
onnx==1.13.1
opencv-python==4.7.0.72
packaging==23.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pipreqs==0.4.12
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
Pygments==2.15.0
python-dateutil==2.8.2
pyzmq==25.0.2
requests==2.28.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
tomli==2.0.1
torch==2.0.0
torchvision==0.15.1
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
urllib3==1.26.15
wcwidth==0.2.6
Werkzeug==2.2.3
wget==3.2
yarg==0.1.9
zipp==3.15.0
